+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/clannad/BeanTransformerConfiguration.java
@@ -1,7 +1,9 @@
 package com.dwarfeng.familyhelper.webapi.node.configuration.clannad;
 
-import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonGenderTypeIndicator;
-import com.dwarfeng.familyhelper.clannad.stack.bean.entity.GenderTypeIndicator;
+import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonProfile;
+import com.dwarfeng.familyhelper.clannad.sdk.bean.entity.FastJsonProfileTypeIndicator;
+import com.dwarfeng.familyhelper.clannad.stack.bean.entity.Profile;
+import com.dwarfeng.familyhelper.clannad.stack.bean.entity.ProfileTypeIndicator;
 import com.dwarfeng.subgrade.impl.bean.DozerBeanTransformer;
 import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
 import org.dozer.Mapper;
@@ -18,8 +20,13 @@ public class BeanTransformerConfiguration {
     }
 
     @Bean
-    public BeanTransformer<GenderTypeIndicator, FastJsonGenderTypeIndicator>
-    genderTypeIndicatorBeanTransformer() {
-        return new DozerBeanTransformer<>(GenderTypeIndicator.class, FastJsonGenderTypeIndicator.class, mapper);
+    public BeanTransformer<Profile, FastJsonProfile> profileBeanTransformer() {
+        return new DozerBeanTransformer<>(Profile.class, FastJsonProfile.class, mapper);
+    }
+
+    @Bean
+    public BeanTransformer<ProfileTypeIndicator, FastJsonProfileTypeIndicator>
+    profileTypeIndicatorBeanTransformer() {
+        return new DozerBeanTransformer<>(ProfileTypeIndicator.class, FastJsonProfileTypeIndicator.class, mapper);
     }
 }