+++ familyhelper-webapi-sdk/src/main/java/com/dwarfeng/familyhelper/webapi/sdk/cna/ValidateList.java
@@ -1,4 +1,4 @@
-package com.dwarfeng.familyhelper.webapi.sdk.util;
+package com.dwarfeng.familyhelper.webapi.sdk.cna;
 
 import javax.validation.Valid;
 import java.util.*;